+++ archive/legacy_tests/dependency_graph_chain_test.go
@@ -8,13 +8,15 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestComplexEntityDependencyGraph(t *testing.T) {
+func TestChainedEntityDependencyGraph(t *testing.T) {
 	// Create test generator
 	g := NewCSVGenerator("test_output", 10, false)
 
-	// Define test entities with a three-way relationship:
-	// User -> GroupMembership <- Group
-	// GroupMembership depends on both User and Group
+	// Define test entities with a chained dependency:
+	// User -> GroupMembership <- Group <- Role
+	// This creates a chain of dependencies where:
+	// - GroupMembership depends on both User and Group
+	// - Role depends on Group
 	entities := map[string]models.Entity{
 		"user1": {
 			DisplayName: "User",
@@ -39,6 +41,25 @@ func TestComplexEntityDependencyGraph(t *testing.T) {
 					ExternalId: "id",
 					Type:       "String",
 					UniqueId:   true,
+				},
+			},
+		},
+		"role1": {
+			DisplayName: "Role",
+			ExternalId:  "Role",
+			Description: "Role entity",
+			Attributes: []models.Attribute{
+				{
+					Name:       "id",
+					ExternalId: "id",
+					Type:       "String",
+					UniqueId:   true,
+				},
+				{
+					Name:       "groupId",
+					ExternalId: "groupId",
+					Type:       "String",
+					UniqueId:   false,
 				},
 			},
 		},
@@ -83,6 +104,12 @@ func TestComplexEntityDependencyGraph(t *testing.T) {
 			FromAttribute: "GroupMembership.groupId",
 			ToAttribute:   "Group.id",
 		},
+		"role_group": {
+			DisplayName:   "role_group",
+			Name:          "role_group",
+			FromAttribute: "Role.groupId",
+			ToAttribute:   "Group.id",
+		},
 	}
 
 	// Build the dependency graph
@@ -97,32 +124,45 @@ func TestComplexEntityDependencyGraph(t *testing.T) {
 	order, err := g.getTopologicalOrder(g.dependencyGraph)
 	require.NoError(t, err, "Failed to get topological order")
 
-	// In this case, both User and Group should come before GroupMembership
-	require.Len(t, order, 3, "Expected 3 entities in topological order")
+	// We expect the following dependencies:
+	// 1. Group should be earlier in the ordering (Role depends on Group)
+	// 2. User has no dependencies, so its position is less constrained
+	// 3. GroupMembership should be after User and Group
+	// 4. Role should be after Group
 
-	// Check if GroupMembership is the last entity in the order
-	// since it depends on both User and Group
+	require.Len(t, order, 4, "Expected 4 entities in topological order")
+
+	// Find the indices of each entity in the order
 	userIndex := -1
 	groupIndex := -1
+	roleIndex := -1
 	membershipIndex := -1
-	
+
 	for i, entityID := range order {
 		switch entityID {
 		case "user1":
 			userIndex = i
 		case "group1":
 			groupIndex = i
+		case "role1":
+			roleIndex = i
 		case "groupMembership1":
 			membershipIndex = i
 		}
 	}
 
-	// Verify that GroupMembership comes after both User and Group
+	// Verify the dependencies:
+	// 1. GroupMembership should come after both User and Group
 	assert.True(t, membershipIndex > userIndex && membershipIndex > groupIndex,
 		"GroupMembership should come after both User and Group in the topological order. "+
-		"User index: %d, Group index: %d, GroupMembership index: %d", 
+			"User index: %d, Group index: %d, GroupMembership index: %d",
 		userIndex, groupIndex, membershipIndex)
 
-	// Verify that we have exactly two edges
-	assert.Len(t, edges, 2, "Expected exactly 2 edges in the graph")
+	// 2. Role should come after Group
+	assert.Greater(t, roleIndex, groupIndex,
+		"Role should come after Group in the topological order. "+
+			"Group index: %d, Role index: %d", groupIndex, roleIndex)
+
+	// Verify that we have exactly three edges
+	assert.Len(t, edges, 3, "Expected exactly 3 edges in the graph")
 }
